+++ database/migrations/2025_03_11_113702_add_deleted_at_column_in_accounting_database_users_table.php
@@ -28,8 +28,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('deleted_at');
-        });
+        if (!Schema::connection($this->connection)->hasColumn('users', 'deleted_at')) {
+            Schema::connection($this->connection)->table('users', function (Blueprint $table) {
+                $table->dropColumn('deleted_at');
+            });
+        }
     }
 };
